refactored POST request added applyToQuest service
