--- .env.example
@@ -1,33 +1,0 @@
-# Database (Neon PostgreSQL)
-DATABASE_URL=postgresql://user:password@host/dbname?sslmode=require
-DATABASE_URL_UNPOOLED=postgresql://user:password@host/dbname?sslmode=require
-
-# Authentication (NextAuth)
-# Generate secret with: openssl rand -base64 32
-NEXTAUTH_URL=http://localhost:3000
-NEXTAUTH_SECRET=your-32-char-secret-here
-
-# Email (SMTP - optional for MVP)
-SMTP_HOST=smtp.gmail.com
-SMTP_PORT=587
-SMTP_USER=your-email@gmail.com
-SMTP_PASS=your-app-password
-ADMIN_EMAIL=admin@adventurersguild.com
-
-# Application
-NEXT_PUBLIC_APP_URL=http://localhost:3000
-
-# DevSync Integration (optional)
-NEXT_PUBLIC_DEVSYNC_API_URL=https://api.devsync.codes
-DEVSYNC_API_KEY=your-devsync-api-key
-
-# Optional: Analytics
-NEXT_PUBLIC_GA_ID=G-XXXXXXXXXX
-
-# Optional: Error Tracking (Sentry)
-SENTRY_DSN=https://xxx@sentry.io/xxx
-NEXT_PUBLIC_SENTRY_DSN=https://xxx@sentry.io/xxx
-
-# Optional: Rate Limiting (Upstash Redis)
-UPSTASH_REDIS_REST_URL=https://your-redis.upstash.io
-UPSTASH_REDIS_REST_TOKEN=your-token-here
