+++ lib/services/types.ts
@@ -1,0 +1,5 @@
+export type ServiceResult<T> = {
+  data: T | null;
+  error: string | null;
+  status: number;
+};
